Prevent defaults for all keypresses
This is a bad solution to use for everything, because things like Ctrl-Shift-J
(developer tools) and Ctrl-Tab (switch tab) no longer work, but it does prevent
the page from scrolling when I press space. In the future, you should be able to
specify which events you want to prevent-default for.