Handle rAF strangeness
requestAnimationFrame does not run when the window is not active. Thus, when
switching away from a tab and then switching back, the next update will be
unnecessarily long, since it accounts for all of the time spent in the different
tab. To remedy this, we keep track of whether or not we just switched away from
the tab, and if so, drop the update.